Extract utils into separate CLI module
Also start testing the various cli functions.
