+++ chalice/cli/utils.py
@@ -1,0 +1,48 @@
+import logging
+
+import botocore.session
+from typing import Any  # noqa
+
+from chalice import __version__ as chalice_version
+
+
+def create_botocore_session(profile=None, debug=False):
+    # type: (str, bool) -> botocore.session.Session
+    session = botocore.session.Session(profile=profile)
+    _add_chalice_user_agent(session)
+    if debug:
+        session.set_debug_logger('')
+        _inject_large_request_body_filter()
+    return session
+
+
+def _add_chalice_user_agent(session):
+    # type: (botocore.session.Session) -> None
+    suffix = '%s/%s' % (session.user_agent_name, session.user_agent_version)
+    session.user_agent_name = 'aws-chalice'
+    session.user_agent_version = chalice_version
+    session.user_agent_extra = suffix
+
+
+def _inject_large_request_body_filter():
+    # type: () -> None
+    log = logging.getLogger('botocore.endpoint')
+    log.addFilter(LargeRequestBodyFilter())
+
+
+class LargeRequestBodyFilter(logging.Filter):
+    def filter(self, record):
+        # type: (Any) -> bool
+        # Note: the proper type should be "logging.LogRecord", but
+        # the typechecker complains about 'Invalid index type "int" for "dict"'
+        # so we're using Any for now.
+        if record.msg.startswith('Making request'):
+            if record.args[0].name in ['UpdateFunctionCode', 'CreateFunction']:
+                # When using the ZipFile argument (which is used in chalice),
+                # the entire deployment package zip is sent as a base64 encoded
+                # string.  We don't want this to clutter the debug logs
+                # so we don't log the request body for lambda operations
+                # that have the ZipFile arg.
+                record.args = (record.args[:-1] +
+                               ('(... omitted from logs due to size ...)',))
+        return True
